requirements.txt
ollama~=0.6.1
pyyaml
httpx